refactor(f): add `constants.ts`
