+++ packages/frontend/src/routes/(app)/about/+page.server.ts
@@ -2,7 +2,7 @@ import { dev } from '$app/environment';
 
 export const csr = dev;
 
-import { site } from '$lib/variables';
+import { site } from '$lib/constants';
 import type { MetaTagsProps } from 'svelte-meta-tags';
 import { createSiteTitle } from '$lib/helpers';
 import type { PageServerLoad } from './$types';
