+++ packages/frontend/src/routes/(app)/archive/tags/[tagName]/+page.server.ts
@@ -1,7 +1,7 @@
 import { error, type ServerLoadEvent } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import type { MetaTagsProps } from 'svelte-meta-tags';
-import { site } from '$lib/variables';
+import { site } from '$lib/constants';
 import type { Article, Tag } from '$lib/schema';
 import { createSiteTitle } from '$lib/helpers';
